chore: add NotificationsBackend_RegisterDevice, make Notifications does not inherit from DbServiceBase
